Simplifying modification of `expected_output.json`
No need to run tests with adjusted launch.json anymore. Just run `utils/set_expected_output.py`
